+++ src/models/Settings.php
@@ -55,7 +55,7 @@ class Settings extends Model
     /**
      * @inheritdoc
      */
-    public function rules()
+    public function rules(): array
     {
         return [
             [['apiPath', 'numberOfFeeds'], 'required'],